Initial Move Calculate the detail metric correctly (on a per-genotype basis). Add more tests Add test to handle multi-sample VCF Refactor detail->summary metrics accumulation. Fix CreateVerifyIDIntensityContaminationMetricsFile to handle LLK and LLK0 as doubles. Added new metric, refined naming of others. Enhanced multiple sample tests.

+++ src/main/java/picard/arrays/VerifyIDIntensityContaminationMetrics.java
@@ -10,8 +10,8 @@ public class VerifyIDIntensityContaminationMetrics extends MetricBase {
     public double PCT_MIX;
 
     /** The log likelihood */
-    public long LLK;
+    public double LLK;
 
     /** The log likelihood 0 */
-    public long LLK0;
+    public double LLK0;
 }
